requirements.txt
Python - 3.8.2
Numpy - 1.21.4
PyTorch - 1.8.1+
Torchvision - 0.9.1+
PIL - 7.1.2
cv2 - 4.6.0
tqdm - 4.46.1